Move getEntityCategory to ProjectAnnotationModel, fix suriano config isEntity
