+++ src/projects/suriano/annotation/ProjectAnnotationModel.ts
@@ -1,7 +1,9 @@
 import {
+  AnnoRepoBody,
   AnnoRepoBodyBase,
   EntityDetail,
 } from "../../../model/AnnoRepoAnnotation.ts";
+import _ from "lodash";
 
 export const projectEntityTypes = ["tf:Ent"];
 export const projectFootnoteMarkerAnnotationTypes = ["tei:Ptr"];
@@ -19,4 +21,8 @@ export function isEntity(
   toTest: AnnoRepoBodyBase,
 ): toTest is ProjectEntityBody {
   return projectEntityTypes.includes(toTest.type);
+}
+
+export function getEntityCategory(annoRepoBody: AnnoRepoBody) {
+  return _.get(annoRepoBody, "metadata.kind") ?? "unknown";
 }
