--- src/projects/suriano/annotation/getEntityCategory.ts
@@ -1,6 +1,0 @@
-import { AnnoRepoBody } from "../../../model/AnnoRepoAnnotation.ts";
-import _ from "lodash";
-
-export function getEntityCategory(annoRepoBody: AnnoRepoBody) {
-  return _.get(annoRepoBody, "metadata.kind") ?? "unknown";
-}
